// backend/src/modules/tasks/entities/enums.ts
export enum PriorityLevel {
  Low = 'baixa',
  Medium = 'media',
  High = 'alta',
  Urgent = 'urgente',
}

export enum Status {
  Backlog = 'pendente',
  ToDo = 'a_fazer',
  InProgress = 'em_andamento',
  Review = 'em_revisao',
  WaitingClient = 'aguardando_cliente',
  Done = 'concluido',
  Cancelled = 'cancelado',
}